feature: :sparkles: #75 initPhaserを別ファイルに切り出す
